requirements.txt
python-dotenv
openai
jinja2
PyYaml
